refactored init
