--- app/actions/index.ts
@@ -1,4 +1,0 @@
-export * from "./trades";
-export * from "./analytics";
-export * from "./psychology";
-export * from "./goals";
