+++ app/page.tsx
@@ -1,7 +1,7 @@
 import Hero from "@/components/hero";
 import ConnectSupabaseSteps from "@/components/tutorial/connect-supabase-steps";
 import SignUpUserSteps from "@/components/tutorial/sign-up-user-steps";
-import { hasEnvVars } from "@/utils/supabase/check-env-vars";
+import { hasEnvVars } from "@/lib/utils/supabase/check-env-vars";
 
 export default async function Index() {
   return (
