+++ app/protected/analytics/error.tsx
@@ -1,0 +1,47 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { RefreshCw } from "lucide-react";
+
+export default function AnalyticsErrorPage({
+  error,
+  reset,
+}: {
+  error: Error & { digest?: string };
+  reset: () => void;
+}) {
+  return (
+    <div className="flex justify-center items-center min-h-[calc(100vh-4rem)] p-4">
+      <Card className="w-full max-w-md">
+        <CardHeader>
+          <CardTitle className="text-destructive">
+            Analytics Data Loading Error
+          </CardTitle>
+          <CardDescription>
+            We encountered an issue retrieving your trading analytics.
+          </CardDescription>
+        </CardHeader>
+        <CardContent className="space-y-4">
+          <p className="text-sm text-muted-foreground">
+            {error.message ||
+              "An unexpected error occurred while loading your analytics."}
+          </p>
+
+          <div className="flex justify-center">
+            <Button onClick={() => reset()} variant="outline" className="gap-2">
+              <RefreshCw className="h-4 w-4" />
+              Try Again
+            </Button>
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
