--- components/LoadingSpinner.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { Loader2 } from "lucide-react";
-
-const LoadingSpinner = ({ message = "Loading..." }) => {
-  return (
-    <div className="flex flex-col items-center justify-center min-h-screen bg-background">
-      <div className="flex flex-col items-center space-y-4">
-        <Loader2 className="h-8 animate-spin text-primary" />
-        <p className="text-sm text-muted-foreground">{message}</p>
-      </div>
-    </div>
-  );
-};
-
-export default LoadingSpinner;
