+++ components/header-auth.tsx
@@ -1,9 +1,9 @@
-import { signOutAction } from "@/app/actions";
-import { hasEnvVars } from "@/utils/supabase/check-env-vars";
+import { signOutAction } from "@/lib/actions/auth";
+import { hasEnvVars } from "@/lib/utils/supabase/check-env-vars";
 import Link from "next/link";
 import { Badge } from "./ui/badge";
 import { Button } from "./ui/button";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/lib/utils/supabase/server";
 
 export default async function AuthButton() {
   const supabase = await createClient();
