--- utils/supabase/check-env-vars.ts
@@ -1,6 +1,0 @@
-// This check can be removed
-// it is just for tutorial purposes
-
-export const hasEnvVars =
-  process.env.NEXT_PUBLIC_SUPABASE_URL &&
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
